Re-design of parsers to allow one-at-a-time message handling
Due to the potential large size of email accounts, more messages can be
present than can be handled at once in memory.  Thus, this version of
the parser (renamed EmailAccount -> formats) separates per-account
information and message-handling, and handles messages via a generator
function so that they don't have to be returned to the caller in one
giant chunk.